# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API Series",
        "version": "1.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000/api/v1"
        },
        {
            "url": "http://animeStore.test/api/v1"
        }
    ],
    "paths": {
...
        "/packages": {
            "get": {
                "tags": [
                    "Packages"
                ],
                "summary": "Obtener los paquetes",
                "description": "Regresa la informacion de los paquetes",
                "operationId": "App\\Http\\Controllers\\PackageController::index",
                "responses": {
                    "200": {
                        "description": "Operacion exitosa"
                    },
                    "401": {
                        "description": "Unauthorized"
                    }
                },
                "security": [
                    {
                        "Bearer Auth": [
                            "write:projects",
                            "read:projects"
                        ]
                    }
                ]
            },
            "post": {
                "tags": [
                    "Packages"
                ],
                "summary": "Añadir un paquete",
                "description": "Añade la informacion perteneciente a un paquete",
                "operationId": "App\\Http\\Controllers\\PackageController::store",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {}
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "OK"
                    },
                    "404": {
                        "description": "Una o más series pertenecientes al paquete\n     *          no se encuentran o el paquete mismo no se encuentra"
                    },
                    "400": {
                        "description": "No hay series en el paquete"
                    },
                    "401": {
                        "description": "Unauthorized"
                    }
                },
                "security": [
                    {
                        "Bearer Auth": [
                            "write:projects",
                            "read:projects"
                        ]
                    }
                ]
            }
        },
        "/packages/{packages}": {
            "get": {
                "tags": [
                    "Packages"
                ],
                "summary": "Obtener un paquete",
                "description": "Regresa la informacion de un paquete",
                "operationId": "App\\Http\\Controllers\\PackageController::show",
                "parameters": [
                    {
                        "name": "packages",
                        "in": "path",
                        "description": "Paquete id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Operacion exitosa"
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "404": {
                        "description": "Serie no encontrada"
                    }
                },
                "security": [
                    {
                        "Bearer Auth": [
                            "write:projects",
                            "read:projects"
                        ]
                    }
                ]
            },
            "put": {
                "tags": [
                    "Packages"
                ],
                "summary": "Actualizar un paquete",
                "description": "Regresa la informacion de un paquete ya actualizada",
                "operationId": "App\\Http\\Controllers\\PackageController::update",
                "parameters": [
                    {
                        "name": "packages",
                        "in": "path",
                        "description": "Paquete id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {}
                        }
                    }
                },
                "responses": {
                    "404": {
                        "description": "Paquete a actualizar no encontrada"
                    },
                    "200": {
                        "description": "Operación exitosa"
                    },
                    "401": {
                        "description": "Unauthorized"
                    }
                },
                "security": [
                    {
                        "Bearer Auth": [
                            "write:projects",
                            "read:projects"
                        ]
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Packages"
                ],
                "summary": "Elimina un paquete",
                "description": "Elimina la información de un paquete",
                "operationId": "App\\Http\\Controllers\\PackageController::destroy",
                "parameters": [
                    {
                        "name": "packages",
                        "in": "path",
                        "description": "Paquete id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Paquete destruido exitosamente"
                    },
                    "404": {
                        "description": "Paquete no encontrado"
                    },
                    "401": {
                        "description": "Unauthorized"
                    }
                },
                "security": [
                    {
                        "Bearer Auth": [
                            "write:projects",
                            "read:projects"
                        ]
                    }
                ]
            }
        },
        "/packages/{packages}/series": {
            "get": {
                "tags": [
                    "Packages"
                ],
                "summary": "Obtener las series pertenecientes a un paquete",
                "description": "Regresa la informacion de las series pertenecientes a un paquete en particular",
                "operationId": "App\\Http\\Controllers\\PackageController::getPackageSeries",
                "parameters": [
                    {
                        "name": "packages",
                        "in": "path",
                        "description": "Paquete id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {}
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Operacion exitosa"
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "404": {
                        "description": "Paquete no encontrada en los registros"
                    }
                },
                "security": [
                    {
                        "Bearer Auth": [
                            "write:projects",
                            "read:projects"
                        ]
                    }
                ]
            },
            "put": {
                "tags": [
                    "Packages"
                ],
                "summary": "Actualizar las series que tiene un paquete",
                "description": "Regresa la información de un paquete,\n     *      con sus series ya actualizadas, de manera que se añaden series al paquete",
                "operationId": "App\\Http\\Controllers\\PackageController::addSeries",
                "parameters": [
                    {
                        "name": "packages",
                        "in": "path",
                        "description": "Paquete id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {}
                        }
                    }
                },
                "responses": {
                    "404": {
                        "description": "Paquete a actualizar no encontrado o una o más series a añadir no encontradas"
                    },
                    "200": {
                        "description": "Operación exitosa"
                    },
                    "401": {
                        "description": "Unauthorized"
                    }
                },
                "security": [
                    {
                        "Bearer Auth": [
                            "write:projects",
                            "read:projects"
                        ]
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Packages"
                ],
                "summary": "Actualizar las series que tiene un paquete",
                "description": "Regresa la información de un paquete,\n     *      con sus series ya actualizadas, de manera que se eliminan series al paquete",
                "operationId": "App\\Http\\Controllers\\PackageController::removeSeries",
                "parameters": [
                    {
                        "name": "packages",
                        "in": "path",
                        "description": "Paquete id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Serie destruida exitosamente"
                    },
                    "404": {
                        "description": "Una o más series no encontradas o paquete no encontrado"
                    },
                    "401": {
                        "description": "Unauthorized"
                    }
                },
                "security": [
                    {
                        "Bearer Auth": [
                            "write:projects",
                            "read:projects"
                        ]
                    }
                ]
            }
        },
        "/series": {
            "get": {
                "tags": [
                    "Series"
                ],
                "summary": "Obtener las series",
                "description": "Regresa la informacion de las series",
                "operationId": "getSeries",
                "responses": {
                    "200": {
                        "description": "Operacion exitosa"
                    },
                    "401": {
                        "description": "Unauthorized"
                    }
                },
                "security": [
                    {
                        "Bearer Auth": [
                            "write:projects",
                            "read:projects"
                        ]
                    }
                ]
            },
            "post": {
                "tags": [
                    "Series"
                ],
                "summary": "Añadir una serie",
                "description": "Añade la informacion perteneciente a una serie",
                "operationId": "App\\Http\\Controllers\\SerieController::store",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {}
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "OK"
                    },
                    "401": {
                        "description": "Unauthorized"
                    }
                },
                "security": [
                    {
                        "Bearer Auth": [
                            "write:projects",
                            "read:projects"
                        ]
                    }
                ]
            }
        },
        "/series/{series}": {
            "get": {
                "tags": [
                    "Series"
                ],
                "summary": "Obtener una serie",
                "description": "Regresa la informacion de una serie",
                "operationId": "getSerie",
                "parameters": [
                    {
                        "name": "series",
                        "in": "path",
                        "description": "Serie id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Operacion exitosa"
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "404": {
                        "description": "Serie no encontrada"
                    }
                },
                "security": [
                    {
                        "Bearer Auth": [
                            "write:projects",
                            "read:projects"
                        ]
                    }
                ]
            },
            "put": {
                "tags": [
                    "Series"
                ],
                "summary": "Actualizar una serie",
                "description": "Regresa la informacion de una serie ya actualizada",
                "operationId": "App\\Http\\Controllers\\SerieController::update",
                "parameters": [
                    {
                        "name": "series",
                        "in": "path",
                        "description": "Serie id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {}
                        }
                    }
                },
                "responses": {
                    "404": {
                        "description": "Serie a actualizar no encontrada"
                    },
                    "200": {
                        "description": "Operación exitosa"
                    },
                    "401": {
                        "description": "Unauthorized"
                    }
                },
                "security": [
                    {
                        "Bearer Auth": [
                            "write:projects",
                            "read:projects"
                        ]
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Series"
                ],
                "summary": "Elimina una serie",
                "description": "Elimina la información de una serie",
                "operationId": "App\\Http\\Controllers\\SerieController::destroy",
                "parameters": [
                    {
                        "name": "series",
                        "in": "path",
                        "description": "Serie id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Serie destruida exitosamente"
                    },
                    "404": {
                        "description": "Serie no encontrada"
                    },
                    "401": {
                        "description": "Unauthorized"
                    }
                },
                "security": [
                    {
                        "Bearer Auth": [
                            "write:projects",
                            "read:projects"
                        ]
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "Bearer Auth": {
                "type": "http",
                "description": "Add bearer token to endpoints",
                "name": "Bearer Token",
                "in": "header",
                "scheme": "bearer"
            }
        }
    }
}